+++ src/main/java/com/kby/home/safety/model/AlertExample.java
@@ -295,52 +295,52 @@ public class AlertExample {
             return (Criteria) this;
         }
 
-        public Criteria andAlertTypeEqualTo(Byte value) {
+        public Criteria andAlertTypeEqualTo(Integer value) {
             addCriterion("alert_type =", value, "alertType");
             return (Criteria) this;
         }
 
-        public Criteria andAlertTypeNotEqualTo(Byte value) {
+        public Criteria andAlertTypeNotEqualTo(Integer value) {
             addCriterion("alert_type <>", value, "alertType");
             return (Criteria) this;
         }
 
-        public Criteria andAlertTypeGreaterThan(Byte value) {
+        public Criteria andAlertTypeGreaterThan(Integer value) {
             addCriterion("alert_type >", value, "alertType");
             return (Criteria) this;
         }
 
-        public Criteria andAlertTypeGreaterThanOrEqualTo(Byte value) {
+        public Criteria andAlertTypeGreaterThanOrEqualTo(Integer value) {
             addCriterion("alert_type >=", value, "alertType");
             return (Criteria) this;
         }
 
-        public Criteria andAlertTypeLessThan(Byte value) {
+        public Criteria andAlertTypeLessThan(Integer value) {
             addCriterion("alert_type <", value, "alertType");
             return (Criteria) this;
         }
 
-        public Criteria andAlertTypeLessThanOrEqualTo(Byte value) {
+        public Criteria andAlertTypeLessThanOrEqualTo(Integer value) {
             addCriterion("alert_type <=", value, "alertType");
             return (Criteria) this;
         }
 
-        public Criteria andAlertTypeIn(List<Byte> values) {
+        public Criteria andAlertTypeIn(List<Integer> values) {
             addCriterion("alert_type in", values, "alertType");
             return (Criteria) this;
         }
 
-        public Criteria andAlertTypeNotIn(List<Byte> values) {
+        public Criteria andAlertTypeNotIn(List<Integer> values) {
             addCriterion("alert_type not in", values, "alertType");
             return (Criteria) this;
         }
 
-        public Criteria andAlertTypeBetween(Byte value1, Byte value2) {
+        public Criteria andAlertTypeBetween(Integer value1, Integer value2) {
             addCriterion("alert_type between", value1, value2, "alertType");
             return (Criteria) this;
         }
 
-        public Criteria andAlertTypeNotBetween(Byte value1, Byte value2) {
+        public Criteria andAlertTypeNotBetween(Integer value1, Integer value2) {
             addCriterion("alert_type not between", value1, value2, "alertType");
             return (Criteria) this;
         }
